+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 
 ### Changed
 
+- Use async io for to speed up retrieval of current resources from GitHub. ([#114](https://gitlab.eclipse.org/eclipsefdn/security/otterdog/-/issues/114))
 - Changed Operation `canonical-diff` to ignore ordering of keys.
 - Support setting a non-existing branch as source branch for GitHub Pages deployment. ([#96](https://gitlab.eclipse.org/eclipsefdn/security/otterdog/-/issues/96))
 - Renamed branch protection rule property `required_approving_reviews` to `requires_pull_request` which is more consistent with its semantics.